pyproject.toml
[project]
name = "port-transparencia-dash"
version = "0.1.0"
description = ""
authors = [
    {name = "Icarobga1610",email = "163910242+Icarobga1610@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "prefect (>=3.4.11,<4.0.0)",
    "pymongo (>=4.13.2,<5.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "pyarrow (>=21.0.0,<22.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
